Add field of view calculation function
This commit adds a new function to calculate the field of view in the horizontal direction. The function takes the focal length and sensor width as inputs and returns the field of view in degrees. This calculation is used to display the field of view in the shooting area component.